+++ atk/atk_test.py
@@ -1,6 +1,7 @@
 import unittest
 
-from . import AlgorithmChain, app
+from atk.atk_algorithm_chain import AtkAlgorithmChain
+from atk import app
 
 
 class AlgorithmTestCase(unittest.TestCase):
@@ -12,7 +13,7 @@ class AlgorithmTestCase(unittest.TestCase):
             'chain_name': 'test_chain',
             'algorithms': []
         }
-        self.cl = AlgorithmChain(path, test_chain).ChainLedger('test')
+        self.cl = AtkAlgorithmChain(path, test_chain).chain_ledger('test')
 
     def check_metadata(self, key, value):
         return self.cl.get_from_metadata(key) == value
